# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CustodianDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianListResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/CustodianDTO"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"data",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"details": {}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianPaginationInfo": {
				"properties": {
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"hasNextPage": {
						"type": "boolean"
					},
					"hasPreviousPage": {
						"type": "boolean"
					}
				},
				"required": [
					"currentPage",
					"itemsPerPage",
					"totalItems",
					"totalPages",
					"hasNextPage",
					"hasPreviousPage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianPaginatedResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"properties": {
							"pagination": {
								"$ref": "#/components/schemas/CustodianPaginationInfo"
							},
							"custodians": {
								"items": {
									"$ref": "#/components/schemas/CustodianDTO"
								},
								"type": "array"
							}
						},
						"required": [
							"pagination",
							"custodians"
						],
						"type": "object"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/CustodianDTO"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianCreateRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Name of the custodian"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianUpdateRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "Updated name of the custodian"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustodianDeleteResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "goldsphere-server",
		"version": "1.0.0",
		"description": "The server including the business logic and integration for the goldsphere project",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/custodians": {
			"get": {
				"operationId": "GetAllCustodians",
				"responses": {
					"200": {
						"description": "List of custodians",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianListResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all custodians (simple list)\nReturns a simple list of all custodians without pagination",
				"tags": [
					"Custodians"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCustodian",
				"responses": {
					"201": {
						"description": "Custodian created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Custodian with this name already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					}
				},
				"description": "Create a new custodian",
				"tags": [
					"Custodians"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Custodian creation data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CustodianCreateRequest",
								"description": "Custodian creation data"
							}
						}
					}
				}
			}
		},
		"/custodians/custodians": {
			"get": {
				"operationId": "GetCustodiansPaginated",
				"responses": {
					"200": {
						"description": "Paginated list of custodians",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CustodianPaginatedResponse"
										},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/CustodianDTO"
													},
													"type": "array"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true
													],
													"nullable": false
												}
											},
											"required": [
												"data",
												"success"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid query parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get custodians with pagination\nRetrieve a paginated list of custodians with filtering and sorting options",
				"tags": [
					"Custodians"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Page number (default: 1)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Items per page (default: 20, max: 100)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Search in custodian names",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort field",
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"custodianName",
								"createdAt",
								"name",
								"updatedAt"
							]
						}
					},
					{
						"description": "Sort order",
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"description": "Filter by active status",
						"in": "query",
						"name": "isActive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Filter by region",
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/custodians/custodians/{id}": {
			"get": {
				"operationId": "GetCustodianById",
				"responses": {
					"200": {
						"description": "Custodian details",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid custodian ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Custodian not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get custodian by ID",
				"tags": [
					"Custodians"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Custodian ID (UUID)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCustodian",
				"responses": {
					"200": {
						"description": "Custodian updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Custodian not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update a custodian",
				"tags": [
					"Custodians"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Custodian ID (UUID)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Updated custodian data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CustodianUpdateRequest",
								"description": "Updated custodian data"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCustodian",
				"responses": {
					"200": {
						"description": "Custodian deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianDeleteResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid custodian ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Custodian not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Cannot delete custodian with existing custody services",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustodianErrorResponse"
								}
							}
						}
					}
				},
				"description": "Delete a custodian",
				"tags": [
					"Custodians"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Custodian ID (UUID)",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	],
...
}